# 2.瑞吉外卖/第二阶段优化/资料/day03/default_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "瑞吉外卖接口文档",
    "version": "1.0",
    "title": "瑞吉外卖"
  },
  "host": "172.17.0.60:8200",
  "basePath": "/",
...
  "paths": {
    "/addressBook": {
      "post": {
        "tags": [
          "地址簿"
        ],
        "summary": "新增",
        "operationId": "saveUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "addressBook",
            "description": "addressBook",
            "required": true,
            "schema": {
              "$ref": "#/definitions/地址管理",
              "originalRef": "地址管理"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«地址管理»",
              "originalRef": "返回值«地址管理»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "地址簿"
        ],
        "summary": "修改",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "addressBook",
            "description": "addressBook",
            "required": true,
            "schema": {
              "$ref": "#/definitions/地址管理",
              "originalRef": "地址管理"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«地址管理»",
              "originalRef": "返回值«地址管理»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "地址簿"
        ],
        "summary": "删除",
        "operationId": "deleteUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "required": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值",
              "originalRef": "返回值"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/addressBook/default": {
      "get": {
        "tags": [
          "地址簿"
        ],
        "summary": "查询默认地址",
        "operationId": "getDefaultUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«地址管理»",
              "originalRef": "返回值«地址管理»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "地址簿"
        ],
        "summary": "默认",
        "operationId": "setDefaultUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "addressBook",
            "description": "addressBook",
            "required": true,
            "schema": {
              "$ref": "#/definitions/地址管理",
              "originalRef": "地址管理"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«地址管理»",
              "originalRef": "返回值«地址管理»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/addressBook/list": {
      "get": {
        "tags": [
          "地址簿"
        ],
        "summary": "列表",
        "operationId": "listUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "cityCode",
            "in": "query",
            "description": "市级区划编号",
            "required": false,
            "type": "string"
          },
          {
            "name": "cityName",
            "in": "query",
            "description": "市级名称",
            "required": false,
            "type": "string"
          },
          {
            "name": "consignee",
            "in": "query",
            "description": "收货人",
            "required": false,
            "type": "string"
          },
          {
            "name": "createTime",
            "in": "query",
            "description": "创建时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "createUser",
            "in": "query",
            "description": "创建人",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "detail",
            "in": "query",
            "description": "详细地址",
            "required": false,
            "type": "string"
          },
          {
            "name": "districtCode",
            "in": "query",
            "description": "区级区划编号",
            "required": false,
            "type": "string"
          },
          {
            "name": "districtName",
            "in": "query",
            "description": "区级名称",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "description": "主键",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "isDefault",
            "in": "query",
            "description": "是否默认 0 否 1是",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "isDeleted",
            "in": "query",
            "description": "是否删除",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "label",
            "in": "query",
            "description": "标签",
            "required": false,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "query",
            "description": "手机号",
            "required": false,
            "type": "string"
          },
          {
            "name": "provinceCode",
            "in": "query",
            "description": "省级区划编号",
            "required": false,
            "type": "string"
          },
          {
            "name": "provinceName",
            "in": "query",
            "description": "省级名称",
            "required": false,
            "type": "string"
          },
          {
            "name": "sex",
            "in": "query",
            "description": "性别 0 女 1 男",
            "required": false,
            "type": "string"
          },
          {
            "name": "updateTime",
            "in": "query",
            "description": "更新时间",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "updateUser",
            "in": "query",
            "description": "修改人",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "userId",
            "in": "query",
            "description": "用户id",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值«List«地址管理»»",
              "originalRef": "返回值«List«地址管理»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/addressBook/{id}": {
      "get": {
        "tags": [
          "地址簿"
        ],
        "summary": "查询",
        "operationId": "getUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/返回值",
              "originalRef": "返回值"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "地址管理": {
      "type": "object",
      "properties": {
        "cityCode": {
          "type": "string",
          "description": "市级区划编号"
        },
        "cityName": {
          "type": "string",
          "description": "市级名称"
        },
        "consignee": {
          "type": "string",
          "description": "收货人"
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间"
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "description": "创建人"
        },
        "detail": {
          "type": "string",
          "description": "详细地址"
        },
        "districtCode": {
          "type": "string",
          "description": "区级区划编号"
        },
        "districtName": {
          "type": "string",
          "description": "区级名称"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "主键"
        },
        "isDefault": {
          "type": "integer",
          "format": "int32",
          "description": "是否默认 0 否 1是"
        },
        "isDeleted": {
          "type": "integer",
          "format": "int32",
          "description": "是否删除"
        },
        "label": {
          "type": "string",
          "description": "标签"
        },
        "phone": {
          "type": "string",
          "description": "手机号"
        },
        "provinceCode": {
          "type": "string",
          "description": "省级区划编号"
        },
        "provinceName": {
          "type": "string",
          "description": "省级名称"
        },
        "sex": {
          "type": "string",
          "description": "性别 0 女 1 男"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间"
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "description": "修改人"
        },
        "userId": {
          "type": "integer",
          "format": "int64",
          "description": "用户id"
        }
      },
      "title": "地址管理"
    },
...
    "返回值": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "编码：1成功，0和其它数字为失败"
        },
        "data": {
          "type": "object",
          "description": "返回对象"
        },
        "map": {
          "type": "object",
          "description": "动态字段"
        },
        "msg": {
          "type": "string",
          "description": "返回信息"
        }
      },
      "title": "返回值"
    },
...
    "返回值«List«地址管理»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "编码：1成功，0和其它数字为失败"
        },
        "data": {
          "type": "array",
          "description": "返回对象",
          "items": {
            "$ref": "#/definitions/地址管理",
            "originalRef": "地址管理"
          }
        },
        "map": {
          "type": "object",
          "description": "动态字段",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string",
          "description": "返回信息"
        }
      },
      "title": "返回值«List«地址管理»»"
    },
...
    "返回值«地址管理»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "编码：1成功，0和其它数字为失败"
        },
        "data": {
          "description": "返回对象",
          "$ref": "#/definitions/地址管理",
          "originalRef": "地址管理"
        },
        "map": {
          "type": "object",
          "description": "动态字段",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string",
          "description": "返回信息"
        }
      },
      "title": "返回值«地址管理»"
    },
...
  }
}